pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.8.0"]

[project]
authors = [
  {name = "Elmano Neto", email = "elmanoneto@gmail.com"},
]
description = "API para análise de sentimentos com FastAPI, VADER e tradução automática"
name = "fastapi-sentiment"
readme = "README.md"
requires-python = ">=3.9,<3.13"
version = "0.1.0"

dependencies = [
  "fastapi>=0.115.2,<0.116.0",
  "uvicorn[standard]>=0.34.2,<0.35.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "textblob>=0.19.0,<0.20.0",
  "vadersentiment>=3.3.2,<4.0.0",
  "deep-translator>=1.11.4,<2.0.0",
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
httpx = "^0.28.1"
isort = "^6.0.1"
pre-commit = "^4.2.0"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[tool.flake8]
max-line-length = 120

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing"
